observable design pattern
