+++ model/Movie/MovieFactory.py
@@ -22,23 +22,3 @@ class MovieFactory:
             movie = StandardMovie.StandardMovie(request.get('title'), request.get('movie_length'), request.get('tickets'))
 
         return movie
-
-
-class Observable():
-    
-    def __init__(self):
-
-        email = pd.read_csv(MOVIE_CSV_PATH_STRING)
-  
-        self.userEmail = email['EMAIL'].tolist()
- 
-        
-    def unregister(self, user):
-        self.userEmail.remove(user.getUserEmail())
-     
-    def notify_user(self):
-        
-        if(addMovieToCSV(Movie)):
-            
-            for user in self.userEmail:
-                user.update(self)
